fix id dokter

+++ src/SimpleCliniq.Module.Core.Domain/Models/MDokter.cs
@@ -7,7 +7,7 @@ namespace SimpleCliniq.Module.Core.Domain.Models;
 
 public partial class MDokter
 {
-    public decimal IdDokter { get; set; }
+    public long IdDokter { get; set; }
 
     public string KdDokter { get; set; }
